--- ledger/ledger-utxo/src/test/java/net/corda/v5/ledger/utxo/IssuableStateJavaApiTests.java
@@ -1,22 +1,0 @@
-package net.corda.v5.ledger.utxo;
-
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-import java.security.PublicKey;
-import java.util.Set;
-
-public final class IssuableStateJavaApiTests extends AbstractMockTestHarness {
-
-    @Test
-    public void getParticipantsShouldReturnTheExpectedValue() {
-        Set<PublicKey> value = issuableState.getParticipants();
-        Assertions.assertEquals(participants, value);
-    }
-
-    @Test
-    public void getIssuerShouldReturnTheExpectedValue() {
-        PublicKey value = issuableState.getIssuer();
-        Assertions.assertEquals(aliceKey, value);
-    }
-}
